+++ projects/ngx-enaio/src/lib/interfaces/enaio-languages.ts
@@ -3,12 +3,18 @@ import { EnaioLanguageStatus } from '../enums/enaio-language-status.enum';
 
 /** Available Languages */
 export interface EnaioLanguages {
-  /** List of all available languages */
-  language: EnaioLanguage[];
+  /** Languages */
+  languages: EnaioLanguage;
 }
 
 /** Enaio Language Info */
 export interface EnaioLanguage {
+  /** Array of all available languages */
+  language: EnaioLanguageItem[];
+}
+
+/** Enaio Language Info */
+export interface EnaioLanguageItem {
   /** Language name */
   name: string;
   /** Language id */